[docker_nginx] use cluster name as dimension for unicorn connections
Thus we can build stack specific alarms on them.
Aggregation dimension was not the best place for this because it is for
the dashboards which should not be changed when deploying.
No need to add additional aggregations for this because the
instance ID/container ID combo is (currently) not needed for those metrics.