socket: read data from existing stats file
Ensures that the socket data reported as metric is consistent with the
data used by the application to block requests - i.e. "single source of
truth".

Keeps the fallback implementation for exceptional conditions:
- the stack was not configured with a stats file
- a relative stats file path is not resolved correctly with the
  hardcoded (assumed) railsapp directory
- the stats file is not accessible from the plugin's container
- the stats file contents could not be parsed successfully
- puma did not yet write the file, because the plugin is run "too early
  for the recent puma startup".

Given `rack_queued` is the important metric, this one is used to decide
whether the fallback needs to be run.